performance(tests) Tweaks to make jest run tests faster.
